backend/pyproject.toml
[project]
name = "robo-agent-backend"
version = "0.1.0"
description = "RoboAgent MVP – AI-powered robotics design assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "sympy>=1.13",
    "numpy>=1.26",
    "chromadb>=0.5",
    "unstructured[pdf]>=0.15",
    "httpx>=0.27",
    "pydantic-settings>=2.4",
    "python-dotenv>=1.0",
    "sse-starlette>=2.1",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"
